Created placeholder opened page +page file
